+++ CHANGELOG.md
@@ -40,6 +40,7 @@ If MD5 sum is not listed for a certain release then it means that the container 
 
 ### Fixed
 
+- Made ``usecases/LDpred2/complementSumstats.R`` write output file by default, not stdout.
 - Fixed print statement in ``usecases/LDpred2/complementSumstats.R`` causing crash w. ``--file-output`` arg.
 - Fixed ``ldpred2.R`` script in case ``--file-pheno``/``--col-pheno``/``--col-pheno-from-fam`` args were used, by removing these options altogether.
 - Use [packagemanager.rstudio.com/cran/__linux__/focal/2023-02-16](https://packagemanager.rstudio.com/cran/__linux__/focal/2023-02-16) as main R package repo